+++ packages/common/src/util/index.js
@@ -1,4 +1,5 @@
 export * from './http';
+export * from './helpers';
 export * from './references';
 import parseDate from './parse-date';